refactor: Update TapTapPage component styling and add submit button
Consolidate the styling of the TapTapPage component by updating the CSS file. This includes adding a new CSS class for the submit button. Additionally, a submit button has been added to the TapTapPage component, allowing users to submit the form. This change improves the overall user experience and functionality of the component.